Moved AlignDetectors v2 to be v1
