+++ Code/Mantid/Framework/Algorithms/inc/MantidAlgorithms/AlignDetectors.h
@@ -2,7 +2,7 @@
 #define MANTID_ALGORITHMS_ALIGNDETECTORS_H_
 
 #include "MantidAPI/Algorithm.h"
-#include "MantidDataObjects/EventWorkspace.h"
+#include "MantidAPI/ITableWorkspace_fwd.h"
 #include "MantidDataObjects/OffsetsWorkspace.h"
 #include "MantidKernel/System.h"
 
@@ -52,23 +52,17 @@ public:
   AlignDetectors();
   virtual ~AlignDetectors();
 
-  /// Algorithm's name for identification overriding a virtual method
-  virtual const std::string name() const { return "AlignDetectors"; };
-  /// Summary of algorithms purpose
-  virtual const std::string summary() const {
-    return "Performs a unit change from TOF to dSpacing, correcting the X "
-           "values to account for small errors in the detector positions.";
-  }
+  /// Algorithms name for identification. @see Algorithm::name
+  virtual const std::string name() const;
+  /// Algorithm's summary for use in the GUI and help. @see Algorithm::summary
+  virtual const std::string summary() const;
 
-  /// Algorithm's version for identification overriding a virtual method
-  virtual int version() const { return 1; };
-  /// Algorithm's category for identification overriding a virtual method
-  virtual const std::string category() const { return "Diffraction"; }
-
-  // ----- Useful static functions ------
-  static std::map<detid_t, double> *calcTofToD_ConversionMap(
-      Mantid::API::MatrixWorkspace_const_sptr inputWS,
-      Mantid::DataObjects::OffsetsWorkspace_sptr offsetsWS);
+  /// Algorithm's version for identification. @see Algorithm::version
+  virtual int version() const ;
+  /// Algorithm's category for identification. @see Algorithm::category
+  virtual const std::string category() const;
+  /// Cross-check properties with each other @see IAlgorithm::validateInputs
+  virtual std::map<std::string, std::string> validateInputs();
 
 private:
   // Implement abstract Algorithm methods
@@ -77,11 +71,14 @@ private:
 
   void execEvent();
 
-  // void execTOFEvent(std::string calfilename,
-  // Mantid::API::MatrixWorkspace_const_sptr inputWS);
+  void loadCalFile(API::MatrixWorkspace_sptr inputWS, const std::string & filename);
+  void getCalibrationWS(API::MatrixWorkspace_sptr inputWS);
 
-  /// Pointer for an event workspace
-  Mantid::DataObjects::EventWorkspace_const_sptr eventW;
+  Mantid::API::ITableWorkspace_sptr m_calibrationWS;
+
+  /// number of spectra in input workspace
+  int64_t m_numberOfSpectra;
+
 
   /// Map of conversion factors for TOF to d-Spacing conversion
   std::map<detid_t, double> *tofToDmap;
