+++ Code/Mantid/Framework/Algorithms/src/AlignDetectors.cpp
@@ -3,6 +3,7 @@
 //----------------------------------------------------------------------
 #include "MantidAlgorithms/AlignDetectors.h"
 #include "MantidAPI/FileProperty.h"
+#include "MantidAPI/ITableWorkspace.h"
 #include "MantidAPI/WorkspaceValidators.h"
 #include "MantidDataObjects/EventWorkspace.h"
 #include "MantidDataObjects/OffsetsWorkspace.h"
@@ -23,91 +24,133 @@ namespace Algorithms {
 // Register the algorithm into the algorithm factory
 DECLARE_ALGORITHM(AlignDetectors)
 
-//-----------------------------------------------------------------------
-/**
- * Make a map of the conversion factors between tof and D-spacing
- * for all pixel IDs in a workspace.
- * @param inputWS the workspace containing the instrument geometry of interest.
- * @param offsetsWS map between pixelID and offset (from the calibration file)
- * @return map of conversion factors between TOF and dSpacing
- */
-std::map<detid_t, double> *AlignDetectors::calcTofToD_ConversionMap(
-    Mantid::API::MatrixWorkspace_const_sptr inputWS,
-    OffsetsWorkspace_sptr offsetsWS) {
-  // Get a pointer to the instrument contained in the workspace
-  Instrument_const_sptr instrument = inputWS->getInstrument();
+namespace { // anonymous namespace
 
-  double l1;
-  Kernel::V3D beamline, samplePos;
-  double beamline_norm;
+    /// Applies the equation d=(TOF-tzero)/difc
+    struct difc_only {
+      difc_only(const double difc) : factor(1./difc){
+      }
 
-  instrument->getInstrumentParameters(l1, beamline, beamline_norm, samplePos);
+      double operator()(const double tof) const {
+        return factor*tof;
+      }
 
-  std::map<detid_t, double> *myMap = new std::map<detid_t, double>();
+      /// 1./difc
+      double factor;
+    };
 
-  // To get all the detector ID's
-  detid2det_map allDetectors;
-  instrument->getDetectors(allDetectors);
+    /// Applies the equation d=(TOF-tzero)/difc
+    struct difc_and_tzero {
+      difc_and_tzero(const double difc, const double tzero) : factor(1./difc), offset(-1.*tzero/difc){
+      }
 
-  // Now go through all
-  detid2det_map::iterator it;
-  for (it = allDetectors.begin(); it != allDetectors.end(); ++it) {
-    detid_t detectorID = it->first;
-    Geometry::IDetector_const_sptr det = it->second;
+      double operator()(const double tof) const {
+        return factor*tof+offset;
+      }
 
-    // Find the offset, if any
-    double offset = offsetsWS->getValue(detectorID, 0.0);
-    if (offset <= -1.) { // non-physical
-      std::stringstream msg;
-      msg << "Encountered offset of " << offset
-          << " which converts data to negative d-spacing for detectorID "
-          << detectorID << "\n";
-      throw std::logic_error(msg.str());
-    }
+      /// 1./difc
+      double factor;
+      /// -tzero/difc
+      double offset;
+    };
 
-    // Compute the factor
-    double factor = Instrument::calcConversion(l1, beamline, beamline_norm,
-                                               samplePos, det, offset);
+    struct difa_positive {
+      difa_positive(const double difc, const double difa, const double tzero) {
+          factor1 = -0.5 * difc / difa;
+          factor2 = 1. / difa;
+          factor3 = (factor1 * factor1) - (tzero / difa);
+      }
 
-    // Save in map
-    (*myMap)[detectorID] = factor;
-  }
+      double operator()(const double tof) const {
+        return factor1 + std::sqrt((tof*factor2) + factor3);
+      }
 
-  // Give back the map.
-  return myMap;
+      /// -0.5*difc/difa
+      double factor1;
+      /// 1/difa
+      double factor2;
+      /// (0.5*difc/difa)^2 - (tzero/difa)
+      double factor3;
+    };
+
+    class ConversionFactors {
+    public:
+        ConversionFactors(ITableWorkspace_const_sptr table) {
+            m_difcCol = table->getColumn("difc");
+            m_difaCol = table->getColumn("difa");
+            m_tzeroCol = table->getColumn("tzero");
+
+            this->generateDetidToRow(table);
+        }
+
+        std::function<double(double)> getConversionFunc(const std::set<detid_t> & detIds) {
+            const std::set<size_t> rows = this->getRow(detIds);
+            double difc = 0.;
+            double difa = 0.;
+            double tzero = 0.;
+            for (auto row = rows.begin(); row != rows.end(); ++row) {
+                difc += m_difcCol->toDouble(*row);
+                difa += m_difaCol->toDouble(*row);
+                tzero += m_tzeroCol->toDouble(*row);
+            }
+            if (rows.size() > 1) {
+                double norm = 1./static_cast<double>(rows.size());
+                difc = norm * difc;
+                difa = norm * difa;
+                tzero = norm * tzero;
+            }
+
+            if (difa == 0.) {
+                if (tzero == 0.) {
+                    return difc_only(difc);
+                } else {
+                    return difc_and_tzero(difc, tzero);
+                }
+            } else if (difa > 0.){
+                return difa_positive(difc, difa, tzero);
+            } else {
+                throw std::runtime_error("difa != 0 branch needs to be written"); // TODO
+            }
+        }
+
+    private:
+        void generateDetidToRow(ITableWorkspace_const_sptr table) {
+            ConstColumnVector<int> detIDs = table->getVector("detid");
+            const size_t numDets = detIDs.size();
+            for (size_t i=0; i<numDets; ++i) {
+                m_detidToRow[static_cast<detid_t>(detIDs[i])]=i;
+            }
+        }
+
+        std::set<size_t> getRow(const std::set<detid_t> & detIds) {
+            std::set<size_t> rows;
+            for (auto detId = detIds.begin(); detId != detIds.end(); ++detId) {
+                auto rowIter = m_detidToRow.find(*detId);
+                if (rowIter != m_detidToRow.end()) { // skip if not found
+                    rows.insert(rowIter->second);
+                }
+            }
+            return rows;
+        }
+
+        std::map<detid_t, size_t> m_detidToRow;
+        Column_const_sptr m_difcCol;
+        Column_const_sptr m_difaCol;
+        Column_const_sptr m_tzeroCol;
+    };
+} // anonymous namespace
+
+const std::string AlignDetectors::name() const { return "AlignDetectors"; }
+
+int AlignDetectors::version() const { return 1; }
+
+const std::string AlignDetectors::category() const { return "Diffraction"; }
+
+const std::string AlignDetectors::summary() const {
+  return "Performs a unit change from TOF to dSpacing, correcting the X "
+         "values to account for small errors in the detector positions.";
 }
 
-//-----------------------------------------------------------------------
-/** Compute a conversion factor for a LIST of detectors.
- * Averages out the conversion factors if there are several.
- *
- * @param tofToDmap :: detectord id to double conversions map
- * @param detectors :: list of detector IDS
- * @return
- */
-
-double calcConversionFromMap(std::map<detid_t, double> *tofToDmap,
-                             const std::set<detid_t> &detectors) {
-  double factor = 0.;
-  detid_t numDetectors = 0;
-  for (std::set<detid_t>::const_iterator iter = detectors.begin();
-       iter != detectors.end(); ++iter) {
-    detid_t detectorID = *iter;
-    std::map<detid_t, double>::iterator it;
-    it = tofToDmap->find(detectorID);
-    if (it != tofToDmap->end()) {
-      factor += it->second; // The factor for that ID
-      numDetectors++;
-    }
-  }
-  if (numDetectors > 0)
-    return factor / static_cast<double>(numDetectors);
-  else
-    return 0;
-}
-
-//========================================================================
-//========================================================================
 /// (Empty) Constructor
 AlignDetectors::AlignDetectors() { this->tofToDmap = NULL; }
 
@@ -120,7 +163,6 @@ void AlignDetectors::init() {
   // Workspace unit must be TOF.
   wsValidator->add<WorkspaceUnitValidator>("TOF");
   wsValidator->add<RawCountValidator>();
-  wsValidator->add<InstrumentValidator>();
 
   declareProperty(new WorkspaceProperty<API::MatrixWorkspace>(
                       "InputWorkspace", "", Direction::Input, wsValidator),
@@ -131,6 +173,7 @@ void AlignDetectors::init() {
                   "The name to use for the output workspace");
 
   std::vector<std::string> exts;
+  // exts.push_back(".hd5"); // TODO
   exts.push_back(".cal");
   declareProperty(
       new FileProperty("CalibrationFile", "", FileProperty::OptionalLoad, exts),
@@ -138,10 +181,106 @@ void AlignDetectors::init() {
       "Either this or OffsetsWorkspace needs to be specified.");
 
   declareProperty(
+      new WorkspaceProperty<ITableWorkspace>(
+          "CalibrationWorkspace", "", Direction::Input, PropertyMode::Optional),
+      "Optional: A Workspace containing the calibration information. Either "
+      "this or CalibrationFile needs to be specified.");
+
+  declareProperty(
       new WorkspaceProperty<OffsetsWorkspace>(
           "OffsetsWorkspace", "", Direction::Input, PropertyMode::Optional),
       "Optional: A OffsetsWorkspace containing the calibration offsets. Either "
       "this or CalibrationFile needs to be specified.");
+
+  // make group associations.
+  std::string calibrationGroup("Calibration");
+  setPropertyGroup("CalibrationFile", calibrationGroup);
+  setPropertyGroup("CalibrationWorkspace", calibrationGroup);
+  setPropertyGroup("OffsetsWorkspace", calibrationGroup);
+}
+
+std::map<std::string, std::string> AlignDetectors::validateInputs() {
+    std::map<std::string, std::string> result;
+
+    int numWays = 0;
+
+    const std::string calFileName = getProperty("CalibrationFile");
+    if (!calFileName.empty()) numWays+=1;
+
+    ITableWorkspace_const_sptr calibrationWS = getProperty("CalibrationWorkspace");
+    if(bool(calibrationWS)) numWays+=1;
+
+    OffsetsWorkspace_const_sptr offsetsWS = getProperty("OffsetsWorkspace");
+    if(bool(offsetsWS)) numWays+=1;
+
+    std::string message;
+    if (numWays == 0) {
+        message = "You must specify only one of CalibrationFile, "
+                  "CalibrationWorkspace, OffsetsWorkspace.";
+    }
+    if (numWays > 1) {
+        message = "You must specify one of CalibrationFile, "
+                  "CalibrationWorkspace, OffsetsWorkspace.";
+    }
+
+    if (!message.empty()) {
+        result["CalibrationFile"] = message;
+        result["CalibrationWorkspace"] = message;
+    }
+
+    return result;
+}
+
+bool endswith(const std::string& str, const std::string &ending) {
+    if (ending.size() > str.size()) {
+      return false;
+    }
+
+    return std::equal(str.begin() + str.size() - ending.size(),
+                      str.end(), ending.begin());
+}
+
+void AlignDetectors::loadCalFile(MatrixWorkspace_sptr inputWS, const std::string & filename) {
+  if (endswith(filename, ".cal")) {
+      // Load the .cal file
+      IAlgorithm_sptr alg = createChildAlgorithm("LoadCalFile");
+      alg->setPropertyValue("CalFilename", filename);
+      alg->setProperty("InputWorkspace", inputWS);
+      alg->setProperty<bool>("MakeGroupingWorkspace", false);
+      alg->setProperty<bool>("MakeOffsetsWorkspace", true);
+      alg->setProperty<bool>("MakeMaskWorkspace", false);
+      alg->setPropertyValue("WorkspaceName", "temp");
+      alg->executeAsChildAlg();
+      m_calibrationWS = alg->getProperty("OutputCalWorkspace");
+//  } else if (endswith(filename, ".hd5")) { // TODO
+
+  } else {
+      std::stringstream msg;
+      msg << "Do not know how to load cal file: " << filename;
+      throw std::runtime_error(msg.str());
+  }
+}
+
+void AlignDetectors::getCalibrationWS(MatrixWorkspace_sptr inputWS) {
+    const std::string calFileName = getPropertyValue("CalibrationFile");
+    if (!calFileName.empty()) {
+      progress(0.0, "Reading calibration file");
+      loadCalFile(inputWS, calFileName);
+    } else {
+      m_calibrationWS = getProperty("CalibrationWorkspace");
+      if (!m_calibrationWS) {
+          OffsetsWorkspace_sptr offsetsWS = getProperty("OffsetsWorkspace");
+          auto alg = createChildAlgorithm("ConvertDiffCal");
+          alg->setProperty("OffsetsWorkspace", offsetsWS);
+          alg->executeAsChildAlg();
+          m_calibrationWS = alg->getProperty("OutputWorkspace");
+          m_calibrationWS->setTitle(offsetsWS->getTitle());
+      }
+    }
+}
+
+void setXAxisUnits(API::MatrixWorkspace_sptr outputWS) {
+    outputWS->getAxis(0)->unit() = UnitFactory::Instance().create("dSpacing");
 }
 
 //-----------------------------------------------------------------------
@@ -155,35 +294,10 @@ void AlignDetectors::exec() {
   // Get the input workspace
   MatrixWorkspace_sptr inputWS = getProperty("InputWorkspace");
 
-  // Read in the calibration data
-  const std::string calFileName = getProperty("CalibrationFile");
-  OffsetsWorkspace_sptr offsetsWS = getProperty("OffsetsWorkspace");
+  this->getCalibrationWS(inputWS);
 
-  progress(0.0, "Reading calibration file");
-  if (offsetsWS && !calFileName.empty())
-    throw std::invalid_argument("You must specify either CalibrationFile or "
-                                "OffsetsWorkspace but not both.");
-  if (!offsetsWS && calFileName.empty())
-    throw std::invalid_argument(
-        "You must specify either CalibrationFile or OffsetsWorkspace.");
-
-  if (!calFileName.empty()) {
-    // Load the .cal file
-    IAlgorithm_sptr alg = createChildAlgorithm("LoadCalFile");
-    alg->setPropertyValue("CalFilename", calFileName);
-    alg->setProperty("InputWorkspace", inputWS);
-    alg->setProperty<bool>("MakeGroupingWorkspace", false);
-    alg->setProperty<bool>("MakeOffsetsWorkspace", true);
-    alg->setProperty<bool>("MakeMaskWorkspace", false);
-    alg->setPropertyValue("WorkspaceName", "temp");
-    alg->executeAsChildAlg();
-    offsetsWS = alg->getProperty("OutputOffsetsWorkspace");
-  }
-
-  const int64_t numberOfSpectra = inputWS->getNumberHistograms();
-
-  // generate map of the tof->d conversion factors
-  this->tofToDmap = calcTofToD_ConversionMap(inputWS, offsetsWS);
+  // Initialise the progress reporting object
+  m_numberOfSpectra = static_cast<int64_t>(inputWS->getNumberHistograms());
 
   // Check if its an event workspace
   EventWorkspace_const_sptr eventW =
@@ -202,20 +316,20 @@ void AlignDetectors::exec() {
   }
 
   // Set the final unit that our output workspace will have
-  outputWS->getAxis(0)->unit() = UnitFactory::Instance().create("dSpacing");
+  setXAxisUnits(outputWS);
 
-  // Initialise the progress reporting object
-  Progress progress(this, 0.0, 1.0, numberOfSpectra);
+  ConversionFactors converter = ConversionFactors(m_calibrationWS);
+
+  Progress progress(this, 0.0, 1.0, m_numberOfSpectra);
 
   // Loop over the histograms (detector spectra)
   PARALLEL_FOR2(inputWS, outputWS)
-  for (int64_t i = 0; i < int64_t(numberOfSpectra); ++i) {
+  for (int64_t i = 0; i < m_numberOfSpectra; ++i) {
     PARALLEL_START_INTERUPT_REGION
     try {
       // Get the input spectrum number at this workspace index
-      const ISpectrum *inSpec = inputWS->getSpectrum(size_t(i));
-      const double factor =
-          calcConversionFromMap(this->tofToDmap, inSpec->getDetectorIDs());
+      auto inSpec = inputWS->getSpectrum(size_t(i));
+      auto toDspacing = converter.getConversionFunc(inSpec->getDetectorIDs());
 
       // Get references to the x data
       MantidVec &xOut = outputWS->dataX(i);
@@ -226,13 +340,13 @@ void AlignDetectors::exec() {
       // vectors were shared.
       const MantidVec &xIn = inSpec->readX();
 
-      // std::transform( xIn.begin(), xIn.end(), xOut.begin(),
-      // std::bind2nd(std::multiplies<double>(), factor) );
+       std::transform( xIn.begin(), xIn.end(), xOut.begin(), toDspacing);
+//       std::bind2nd(std::multiplies<double>(), factor) );
       // the above transform creates wrong output in parallel in debug in Visual
       // Studio
-      for (size_t k = 0; k < xOut.size(); ++k) {
-        xOut[k] = xIn[k] * factor;
-      }
+//      for (size_t k = 0; k < xOut.size(); ++k) {
+//        xOut[k] = xIn[k] * factor;
+//      }
 
       // Copy the Y&E data
       outputWS->dataY(i) = inSpec->readY();
@@ -288,23 +402,18 @@ void AlignDetectors::execEvent() {
   }
 
   // Set the final unit that our output workspace will have
-  outputWS->getAxis(0)->unit() = UnitFactory::Instance().create("dSpacing");
+  setXAxisUnits(outputWS);
 
-  const int64_t numberOfSpectra =
-      static_cast<int64_t>(inputWS->getNumberHistograms());
+  ConversionFactors converter = ConversionFactors(m_calibrationWS);
 
-  // Initialise the progress reporting object
-  Progress progress(this, 0.0, 1.0, numberOfSpectra);
+  Progress progress(this, 0.0, 1.0, m_numberOfSpectra);
 
   PARALLEL_FOR_NO_WSP_CHECK()
-  for (int64_t i = 0; i < int64_t(numberOfSpectra); ++i) {
+  for (int64_t i = 0; i < m_numberOfSpectra; ++i) {
     PARALLEL_START_INTERUPT_REGION
-    // Compute the conversion factor
-    double factor = calcConversionFromMap(
-        this->tofToDmap, inputWS->getSpectrum(size_t(i))->getDetectorIDs());
 
-    // Perform the multiplication on all events
-    outputWS->getEventList(i).convertTof(factor);
+    auto toDspacing = converter.getConversionFunc(inputWS->getSpectrum(size_t(i))->getDetectorIDs());
+    outputWS->getEventList(i).convertTof(toDspacing);
 
     progress.report();
     PARALLEL_END_INTERUPT_REGION
